requirements.txt
opencv-python==4.10.0.84
pillow==10.4.0
pandas==2.2.2
numpy==1.26.4
requests==2.32.2
